refactor: :recycle: Move component codes to dedicated files
