+++ src/useClickAway/useClickAway.ts
@@ -1,0 +1,21 @@
+import React from 'react'
+
+export const useClickAway = (ref: React.RefObject<Element>, callback?: (e: Event) => void) => {
+  React.useEffect(() => {
+    const listener = (e: Event) => {
+      if (!ref.current || ref.current.contains(e.target as Node)) {
+        return
+      }
+
+      callback?.(e)
+    }
+
+    document.addEventListener('mousedown', listener)
+    document.addEventListener('touchstart', listener)
+
+    return () => {
+      document.removeEventListener('mousedown', listener)
+      document.removeEventListener('touchstart', listener)
+    }
+  }, [ref, callback])
+}
