+++ src/useDate/index.ts
@@ -1,0 +1,1 @@
+export { useDate } from './useDate'
