+++ src/useDatepicker/useDatepicker.spec.tsx
@@ -2,7 +2,7 @@ import { fireEvent, render } from '@testing-library/react'
 import { endOfMonth, endOfWeek, format, startOfMonth, startOfWeek } from 'date-fns'
 import React from 'react'
 import { createEvent } from '../../testing/utils'
-import { useDatepicker, UseDatepickerOptions } from './index'
+import { useDatepicker, UseDatepickerOptions } from './useDatepicker'
 
 const Comp: React.FC<Partial<UseDatepickerOptions>> = (props) => {
   const {
