+++ src/useKeyUp/useKeyUp.spec.ts
@@ -1,7 +1,7 @@
 import { fireEvent } from '@testing-library/react'
 import { renderHook, act } from '@testing-library/react-hooks'
 import { createEvent } from '../../testing/utils'
-import { useKeyUp } from './index'
+import { useKeyUp } from './useKeyUp'
 
 it('callback should be called when Esc key keyup.', () => {
   const fn = jest.fn()
