+++ src/useToggle/useToggle.ts
@@ -1,0 +1,17 @@
+import React from 'react'
+
+export const useToggle = (initialState: boolean): [boolean, (state?: unknown) => void] => {
+  const [state, setState] = React.useState(initialState)
+  const toggle = React.useCallback(
+    (nextState?: unknown) =>
+      setState((state) => {
+        if (typeof nextState === 'boolean') {
+          return nextState
+        } else {
+          return !state
+        }
+      }),
+    []
+  )
+  return [state, toggle]
+}
